move link into joint
